refactor(browser): extract Google auth handoff to stay under 1k-line budget
Move shared-cookie login-window logic into side_browser_google_auth.rs so
side_browser_host.rs stays under the FILES_OVER_1K_BUDGET gate.
